requirements.txt
djangorestframework
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.7.7 --no-binary psycopg2
pytz==2018.9
whitenoise==4.1.2
